+++ llvm-external-projects/iree-dialects/include/iree-dialects/Dialect/LinalgExt/Passes/Passes.h
@@ -21,6 +21,8 @@ std::unique_ptr<OperationPass<func::FuncOp>> createLinalgExtToLoopsPass();
 
 std::unique_ptr<OperationPass<>> createPadContractionToBlockSizePass();
 
+std::unique_ptr<OperationPass<func::FuncOp>> createTopkSplitReductionPass();
+
 void registerTilingInterfaceExternalModels(DialectRegistry &registry);
 
 void registerPasses();
