+++ llvm-external-projects/iree-dialects/lib/Dialect/LinalgExt/Passes/CMakeLists.txt
@@ -2,6 +2,7 @@ add_mlir_library(IREELinalgExtPasses
   ConvertToLoops.cpp
   PadContractionToBlockSize.cpp
   Passes.cpp
+  SplitReduction.cpp
   Tiling.cpp
 
   DEPENDS
