+++ llvm-external-projects/iree-dialects/lib/Dialect/LinalgExt/Transforms/CMakeLists.txt
@@ -3,7 +3,6 @@ add_mlir_library(IREELinalgExtTransforms
   InParallelToAsync.cpp
   InParallelToHAL.cpp
   InParallelToSequentialFor.cpp
-  SplitReduction.cpp
   TilingExternalModels.cpp
   TileToSequentialFor.cpp
   TileToInParallel.cpp
