+++ llvm-external-projects/iree-dialects/test/lib/Dialect/CMakeLists.txt
@@ -1,2 +1,1 @@
-add_subdirectory(LinalgExt)
 add_subdirectory(LinalgTransform)
