+++ llvm-external-projects/iree-dialects/tools/iree-dialects-opt/CMakeLists.txt
@@ -9,7 +9,6 @@ set(LIBS
   IREELinalgTransformDialect
   IREELinalgTransformDialectPasses
   IREELinalgTransformTestPasses
-  IREELinalgExtTransformTestPasses
   IREETransformsTestPasses
   IREEPyDMDialect
   IREEPyDMPasses
